Implement templates for denial request

+++ src/Controller/RequestHandlerController.php
@@ -3,10 +3,12 @@
 namespace App\Controller;
 
 use App\Entity\AccessRequest;
+use App\Form\AccessRequestDenyFormType;
 use App\Repository\AccessRequestRepository;
 use App\Service\Mail\MailHelper;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
@@ -68,11 +70,12 @@ class RequestHandlerController extends AbstractController
      * This will send an e-mail to the user notifying them about the denial.
      * If entered by the cluster administrator, a brief explanation will be added to the mail.
      *
-     * @Route("/deny/{id}", name="deny_request", methods={"GET"})
+     * @Route("/deny/{id}", name="deny_request", methods={"GET", "POST"})
      * @param string $id
+     * @param Request $request
      * @return Response
      */
-    public function denyRequest(string $id): Response
+    public function denyRequest(string $id, Request $request): Response
     {
         /** @var AccessRequest $accessRequest */
         $accessRequest = $this->requestRepository->find($id);
@@ -81,13 +84,23 @@ class RequestHandlerController extends AbstractController
         }
 
         // TODO: check form submit for reason. Proceed with mail send on submit
+        $form = $this->createForm(AccessRequestDenyFormType::class);
+        $form->handleRequest($request);
 
-        $this->mailHelper->sendDeniedMail($accessRequest);
+        if ($form->isSubmitted() && $form->isValid()) {
+            $data = $form->getData();
+            $reason = $data['reason'] ?: null;
+            $this->mailHelper->sendDeniedMail($accessRequest, $reason);
 
-        $this->em->remove($accessRequest);
-        $this->em->flush();
+            $this->em->remove($accessRequest);
+            $this->em->flush();
 
-        return $this->render('request/request_denied.html.twig');
+            return $this->render('request/request_denied.html.twig');
+        }
+
+        return $this->render('request/request_deny.html.twig', [
+            'form' => $form->createView()
+        ]);
     }
 
 }
